+++ scheduler/__init__.py
@@ -4,7 +4,7 @@ A simple in-process python `scheduler` library for `datetime` objects.
 Author: Jendrik A. Potyka, Fabian A. Preiss
 """
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 __author__ = "Jendrik A. Potyka, Fabian A. Preiss"
 
 from scheduler.core import Scheduler